One class per file, remove nested fixture
